editors - adopt working copy handler for untitled
