+++ src/vs/workbench/services/workingCopy/common/legacyBackupRestorer.ts
@@ -9,10 +9,8 @@ import { IEditorService } from 'vs/workbench/services/editor/common/editorServic
 import { IResourceEditorInput } from 'vs/platform/editor/common/editor';
 import { Schemas } from 'vs/base/common/network';
 import { ILifecycleService, LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle';
-import { IUntitledTextResourceEditorInput, IEditorInput, IEditorInputWithOptions } from 'vs/workbench/common/editor';
-import { toLocalResource, isEqual } from 'vs/base/common/resources';
-import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
-import { IPathService } from 'vs/workbench/services/path/common/pathService';
+import { IEditorInput } from 'vs/workbench/common/editor';
+import { isEqual } from 'vs/base/common/resources';
 import { ILogService } from 'vs/platform/log/common/log';
 import { Promises } from 'vs/base/common/async';
 import { IWorkingCopyIdentifier } from 'vs/workbench/services/workingCopy/common/workingCopy';
@@ -22,14 +20,10 @@ import { IWorkingCopyIdentifier } from 'vs/workbench/services/workingCopy/common
  */
 export class LegacyWorkingCopyBackupRestorer implements IWorkbenchContribution {
 
-	private static readonly UNTITLED_REGEX = /Untitled-\d+/;
-
 	constructor(
 		@IEditorService private readonly editorService: IEditorService,
 		@IWorkingCopyBackupService private readonly workingCopyBackupService: IWorkingCopyBackupService,
 		@ILifecycleService private readonly lifecycleService: ILifecycleService,
-		@IWorkbenchEnvironmentService private readonly environmentService: IWorkbenchEnvironmentService,
-		@IPathService private readonly pathService: IPathService,
 		@ILogService private readonly logService: ILogService
 	) {
 		this.restoreLegacyBackups();
@@ -46,11 +40,12 @@ export class LegacyWorkingCopyBackupRestorer implements IWorkbenchContribution {
 		// - any working copy without `typeId`
 		// - not `search-edior:/` (supports migration to typeId)
 		const backups = (await this.workingCopyBackupService.getBackups())
-			.filter(backup => backup.typeId.length === 0) 		// any working copy with `typeId` is adopted
+			.filter(backup => backup.typeId.length === 0) 					// any working copy with `typeId` is adopted
 			.filter(backup =>
-				backup.resource.scheme !== 'search-editor' && 			// search editor has adopted
-				backup.resource.scheme !== Schemas.vscodeNotebook &&	// notebooks (complex) has adopted
-				backup.resource.scheme !== Schemas.vscodeCustomEditor	// custom editors (complex) has adopted
+				backup.resource.scheme !== 'search-editor' && 				// search editor has adopted
+				backup.resource.scheme !== Schemas.vscodeNotebook &&		// notebooks (complex) has adopted
+				backup.resource.scheme !== Schemas.vscodeCustomEditor &&	// custom editors (complex) has adopted
+				backup.resource.scheme !== Schemas.untitled					// untitled text editors has adopted
 			);
 
 		// Trigger `resolve` in each opened editor that can be found
@@ -107,19 +102,14 @@ export class LegacyWorkingCopyBackupRestorer implements IWorkbenchContribution {
 		await this.editorService.openEditors(editors);
 	}
 
-	private async resolveEditor(backup: IWorkingCopyIdentifier, index: number, hasOpenedEditors: boolean): Promise<IResourceEditorInput | IUntitledTextResourceEditorInput | IEditorInputWithOptions> {
-
-		// Set editor as `inactive` if we have other editors
-		const options = { pinned: true, preserveFocus: true, inactive: index > 0 || hasOpenedEditors };
-
-		// This is a (weak) strategy to find out if the untitled input had
-		// an associated file path or not by just looking at the path. and
-		// if so, we must ensure to restore the local resource it had.
-		if (backup.resource.scheme === Schemas.untitled && !LegacyWorkingCopyBackupRestorer.UNTITLED_REGEX.test(backup.resource.path)) {
-			return { resource: toLocalResource(backup.resource, this.environmentService.remoteAuthority, this.pathService.defaultUriScheme), options, forceUntitled: true };
-		}
-
-		// Finally return with a simple resource based input
-		return { resource: backup.resource, options };
+	private async resolveEditor(backup: IWorkingCopyIdentifier, index: number, hasOpenedEditors: boolean): Promise<IResourceEditorInput> {
+		return {
+			resource: backup.resource,
+			options: {
+				pinned: true,
+				preserveFocus: true,
+				inactive: index > 0 || hasOpenedEditors // Set editor as `inactive` if we have other editors
+			}
+		};
 	}
 }
